pyproject.toml
[project]
name = "vibe-trade-api"
version = "0.1.0"
description = "API server for vibe-trade with JWT authentication"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "google-cloud-firestore>=2.14.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.12",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

